Cleaned up Index module
