+++ src/app/index/update/inserter.rs
