+++ src/app/index/update/mod.rs
@@ -1,0 +1,75 @@
+use anyhow::*;
+use diesel;
+use diesel::prelude::*;
+use log::{error, info};
+use regex::Regex;
+use std::time;
+
+use crate::config::MiscSettings;
+use crate::db::misc_settings;
+
+mod cleaner;
+mod collector;
+mod inserter;
+mod traverser;
+
+use super::*;
+use cleaner::Cleaner;
+use collector::Collector;
+use inserter::Inserter;
+use traverser::Traverser;
+
+impl Index {
+	pub fn update(&self) -> Result<()> {
+		let start = time::Instant::now();
+		info!("Beginning library index update");
+
+		let album_art_pattern = {
+			let connection = self.db.connect()?;
+			let settings: MiscSettings = misc_settings::table.get_result(&connection)?;
+			Regex::new(&settings.index_album_art_pattern)?
+		};
+
+		let cleaner = Cleaner::new(self.db.clone(), self.vfs_manager.clone());
+		cleaner.clean()?;
+
+		let (insert_sender, insert_receiver) = crossbeam_channel::unbounded();
+		let inserter_db = self.db.clone();
+		let insertion_thread = std::thread::spawn(move || {
+			let mut inserter = Inserter::new(inserter_db, insert_receiver);
+			inserter.insert();
+		});
+
+		let (collect_sender, collect_receiver) = crossbeam_channel::unbounded();
+		let collector_thread = std::thread::spawn(move || {
+			let collector = Collector::new(collect_receiver, insert_sender, album_art_pattern);
+			collector.collect();
+		});
+
+		let vfs = self.vfs_manager.get_vfs()?;
+		let traverser_thread = std::thread::spawn(move || {
+			let mount_points = vfs.get_mount_points();
+			let traverser = Traverser::new(collect_sender);
+			traverser.traverse(mount_points.values().map(|p| p.clone()).collect());
+		});
+
+		if let Err(e) = traverser_thread.join() {
+			error!("Error joining on traverser thread: {:?}", e);
+		}
+
+		if let Err(e) = collector_thread.join() {
+			error!("Error joining on collector thread: {:?}", e);
+		}
+
+		if let Err(e) = insertion_thread.join() {
+			error!("Error joining on inserter thread: {:?}", e);
+		}
+
+		info!(
+			"Library index update took {} seconds",
+			start.elapsed().as_millis() as f32 / 1000.0
+		);
+
+		Ok(())
+	}
+}
