+++ src/app/mod.rs
@@ -1,4 +1,5 @@
 pub mod ddns;
+pub mod index;
 pub mod playlists;
 pub mod thumbnails;
 pub mod vfs;
